refactor: Extract services layer (Part 2 of modularization)
Completed extraction of business logic from mcp_server.py:
- Created services/content_scraper.py with scrape_search_result() function
- Created services/search_processor.py with process_search_results() function
- Updated all import paths to use relative imports (models.*, clients.*, services.*)
- Removed duplicate function definitions from mcp_server.py
- Cleaned up unnecessary imports (json, html2text, requests, BeautifulSoup, Document)

Impact:
- mcp_server.py reduced from 496 to 208 lines (58% reduction)
- Better separation of concerns: scraping, processing, and API coordination
- Improved testability with isolated service functions

Files changed:
- docker/services/content_scraper.py (NEW) - 166 lines
- docker/services/search_processor.py (NEW) - 35 lines
- docker/mcp_server.py (MODIFIED) - 208 lines (was 496)
- docker/clients/*.py (MODIFIED) - Updated imports
- docker/models/*.py (MODIFIED) - Updated imports

Part 2 of 3: Services layer complete. Next: Extract MCP tools layer.
